refactored taskMenu state to separate reducer
