+++ src/modules/AuthListener/index.js
@@ -4,12 +4,10 @@ import store from "../localStore/store"
 import {updateAuthStatusAction} from "../localStore/actions/authActions";
 
 export default class AuthListener {
-
     static init() {
         firebase.auth().onAuthStateChanged(user => {
             const action = updateAuthStatusAction(!!user)
             store.dispatch(action)
-
         })
     }
 }
