+++ src/modules/localStore/store.js
@@ -1,5 +1,10 @@
-import {createStore} from "redux"
-import rootReducer  from "./reducers/reducer"
+import {combineReducers, createStore} from "redux"
+import rootReducer  from "./reducers"
+import authReducer from "./reducers/authReducer";
+
+combineReducers(rootReducer,authReducer)
+
+
 
 const store = createStore(rootReducer)
 export default store
